Move assignment types to primitives
