+++ runtime/parachains/src/inclusion.rs
@@ -25,7 +25,7 @@ use primitives::{
 	parachain::{
 		ValidatorId, AbridgedCandidateReceipt, ValidatorIndex, Id as ParaId,
 		AvailabilityBitfield as AvailabilityBitfield, SignedAvailabilityBitfields, SigningContext,
-		BackedCandidate,
+		BackedCandidate, CoreIndex, GroupIndex, CoreAssignment,
 	},
 };
 use frame_support::{
@@ -38,7 +38,7 @@ use bitvec::{order::Lsb0 as BitOrderLsb0, vec::BitVec};
 use sp_staking::SessionIndex;
 use sp_runtime::{DispatchError, traits::{One, Saturating}};
 
-use crate::{configuration, paras, scheduler::{CoreIndex, GroupIndex, CoreAssignment}};
+use crate::{configuration, paras};
 
 /// A bitfield signed by a validator indicating that it is keeping its piece of the erasure-coding
 /// for any backed candidates referred to by a `1` bit available.
